requirements.txt
azure-cognitiveservices-speech
ffmpeg
moviepy
requests
setuptools